+++ pkg/data/measurement/formatindextag.go
@@ -6,7 +6,7 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/Sirupsen/logrus"
+	"github.com/sirupsen/logrus"
 )
 
 // two byte-oriented functions identical except for operator comparing c to 127.